bridge/requirements.txt
fastapi
uvicorn
gitpython
requests
